Added basic confirmation dialog before deleting entry

+++ web/app/src/app/work/work.component.ts
@@ -59,7 +59,9 @@ export class WorkComponent implements OnInit {
   }
 
   public deleteEntry(id: number): void {
-    this.es.deleteEntry(id);
+    if(window.confirm("Are you sure you want to delete this entry?")) {
+      this.es.deleteEntry(id);
+    }
   }
 
 }
